fix seeders + refactor

+++ api/app/Http/Requests/GeolocationRequest.php
@@ -11,8 +11,8 @@ class GeolocationRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'latitude' => ['required', 'float', 'between:-90,90'],
-            'longitude' => ['required', 'float', 'between:-180,180'],
+            'latitude' => ['required', 'numeric', 'between:-90,90'],
+            'longitude' => ['required', 'numeric', 'between:-180,180'],
         ];
     }
 }
